@@ -1,1 +1,41 @@
-{"name":"ng2-uselect","version":"1.4.5","description":"Universal select component with sortable","main":"bundles/ng2-uselect.umd.js","module":"ng2-uselect.es5.js","es2015":"ng2-uselect.js","typings":"ng2-uselect.d.ts","scripts":{"test":"ng test -sr","clean":"rimraf dist","build":"rimraf dist && ng-packagr -p ng-package.json"},"repository":{"type":"git","url":"git@gitlab.anocms.org:angular/ng2-uselect.git"},"keywords":["select","angular2"],"author":"Alex Anzelm <anex.work@gmail.com>","license":"MIT","dependencies":{"lodash":"^4.17.4"},"peerDependencies":{"@angular/common":"^5.2.7","@angular/core":"^5.2.7","lodash":"^4.17.5"},"devDependencies":{"@angular/common":"^5.2.7","@angular/core":"^5.2.7","@angular/forms":"^5.2.7","bootstrap":"^4.0.0","lodash":"^4.17.5","ng-packagr":"^1.6.0","rxjs":"^5.5.6","typescript":"^2.7.2"},"metadata":"ng2-uselect.metadata.json"}
+{
+  "name": "ng2-uselect",
+  "version": "1.5.0",
+  "description": "Universal select component with sortable",
+  "main": "bundles/ng2-uselect.umd.js",
+  "module": "esm5/ng2-uselect.js",
+  "es2015": "esm2015/ng2-uselect.js",
+  "typings": "ng2-uselect.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git@gitlab.anocms.org:angular/ng2-uselect.git"
+  },
+  "keywords": [
+    "select",
+    "angular2"
+  ],
+  "author": "Alex Anzelm <anex.work@gmail.com>",
+  "license": "MIT",
+  "dependencies": {
+    "@angular/compiler": "^6.0.2",
+    "@angular/compiler-cli": "^6.0.2",
+    "lodash": "^4.17.5",
+    "tslib": "^1.9.0"
+  },
+  "peerDependencies": {
+    "@angular/common": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "lodash": "^4.17.5"
+  },
+  "devDependencies": {
+    "@angular/common": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/forms": "^6.0.2",
+    "bootstrap": "^4.0.0",
+    "lodash": "^4.17.5",
+    "ng-packagr": "^2.4.4",
+    "rxjs": "^6.1.0",
+    "typescript": "^2.7.2"
+  },
+  "metadata": "ng2-uselect.metadata.json"
+}
