@@ -1,6 +1,6 @@
 {
   "name": "ng2-uselect",
-  "version": "1.4.5",
+  "version": "1.5.0",
   "description": "Universal select component with sortable",
   "main": "./dist/bundles/ng2-uselect.umd.js",
   "module": "./dist/ng2-uselect.es5.js",
@@ -19,21 +19,23 @@
   "author": "Alex Anzelm <anex.work@gmail.com>",
   "license": "MIT",
   "dependencies": {
-    "lodash": "^4.17.4"
+    "@angular/compiler": "^6.0.2",
+    "@angular/compiler-cli": "^6.0.2",
+    "lodash": "^4.17.5"
   },
   "peerDependencies": {
-    "@angular/common": "^5.2.7",
-    "@angular/core": "^5.2.7",
+    "@angular/common": "^6.0.2",
+    "@angular/core": "^6.0.2",
     "lodash": "^4.17.5"
   },
   "devDependencies": {
-    "@angular/common": "^5.2.7",
-    "@angular/core": "^5.2.7",
-    "@angular/forms": "^5.2.7",
+    "@angular/common": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "@angular/forms": "^6.0.2",
     "bootstrap": "^4.0.0",
     "lodash": "^4.17.5",
-    "ng-packagr": "^1.6.0",
-    "rxjs": "^5.5.6",
+    "ng-packagr": "^2.4.4",
+    "rxjs": "^6.1.0",
     "typescript": "^2.7.2"
   }
 }
