@@ -1,8 +1,8 @@
 {
   "name": "ng2-uselect",
   "peerDependencies": {
-    "@angular/common": "^5.1.3",
-    "@angular/core": "^5.1.3",
-    "lodash": "^4.17.4"
+    "@angular/common": "^6.0.2",
+    "@angular/core": "^6.0.2",
+    "lodash": "^4.17.5"
   }
 }
